src/main/java/nextstep/subway/auth/ui/session/SessionAuthenticationConverter.java
package nextstep.subway.auth.ui.session;

import nextstep.subway.auth.domain.AuthenticationToken;
import nextstep.subway.auth.ui.AuthenticationConverter;

import javax.servlet.http.HttpServletRequest;
import java.util.Map;

public class SessionAuthenticationConverter implements AuthenticationConverter {

    private static final String USERNAME_FIELD = "username";
    private static final String PASSWORD_FIELD = "password";

    @Override
    public AuthenticationToken convert(HttpServletRequest request) {
        Map<String, String[]> paramMap = request.getParameterMap();
        String principal = paramMap.get(USERNAME_FIELD)[0];
        String credential = paramMap.get(PASSWORD_FIELD)[0];

        return new AuthenticationToken(principal, credential);
    }
}
